Finally fix that one bug
Fix bug where an apple spawning in the tile just after the end of
thesnake's tail would cause the apple to be invisible. This was caused
by snake_remove() and is_snake() being one tile off (with snake_remove()
checking one tile past the end of the snake). snake_remove() was more
accurate to the internal represenation in the calculator's memory, so
is_snake() was updated to check one segment further, and the visual
represenation of the snake was updated by drawing the snake at
thebeginning of the game loop. I can rest easy now.

Signed-off-by: Casey Avila <casey@theavilas.org>